requirements.txt
# --- Web & API ---
fastapi==0.104.1
uvicorn==0.24.0
streamlit>=1.32.0

pandas>=2.1.0
plotly==5.17.0
altair>=5.0.0
streamlit-cookies-controller==0.0.4
apify-client
bcrypt

# --- Database & Backend ---
supabase==2.0.3
python-dotenv==1.0.0

# --- AI & Deep Learning ---
requests>=2.28.0
torch==2.0.1+cpu  --index-url https://download.pytorch.org/whl/cpu
torch-scatter==2.1.2+pt20cpu  -f https://data.pyg.org/whl/torch-2.0.1+cpu.html
torch_sparse
torch_cluster
torch_spline_conv
torch_geometric==2.3.1
transformers>=4.30.0
sentencepiece>=0.1.99

# --- Data & Math ---
numpy>=1.24.0
scikit-learn>=1.3.0
networkx>=3.1

# --- Visualization ---
matplotlib>=3.7.0
beautifulsoup4==4.12.3

# --- Utilities ---
Pillow>=10.0.0